Fix incompatibility with menu swapper shift click
I think the problem was with the menu swapper looking at the current
item id in the clicked item. Since we change the item id to display
the future item, the plugin miscalculated what to do somehow.

The fix is to to the equipping at the end of the frame to allow
menu swapper to behave correctly.